server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
sqlalchemy>=2.0.36
pillow>=11.0.0
piexif>=1.1.3
